refactor: move row/value concepts into module
