+++ read_buffer/src/lib.rs
@@ -7,6 +7,7 @@ pub(crate) mod column;
 pub(crate) mod row_group;
 mod schema;
 pub(crate) mod table;
+pub(crate) mod value;
 
 use std::{
     collections::{btree_map::Entry, BTreeMap, BTreeSet},
@@ -776,9 +777,9 @@ mod test {
         },
         datatypes::DataType::{Boolean, Float64, Int64, UInt64},
     };
-
-    use column::Values;
     use data_types::schema::builder::SchemaBuilder;
+
+    use crate::value::Values;
 
     // helper to make the `database_update_chunk` test simpler to read.
     fn gen_recordbatch() -> RecordBatch {
